Решение ДЗ-5 - разбиение на модули
