+++ index.html
@@ -235,8 +235,8 @@
     </section>
   </template>
 
-  <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <!--<script src="js/functions.js"></script> -->
+  <script src="js/main.js" type="module"></script>
 </body>
 
 </html>
