+++ js/edit-images.js
@@ -1,0 +1,1 @@
+// Модуль редактирования изображений
