+++ js/filtering-images.js
@@ -1,0 +1,1 @@
+// Модуль фильтации изображений
