+++ js/functions.js
@@ -1,3 +1,5 @@
+// Модуль для решения дополнительных учебгых заданий
+
 // Домашнее задание к разделу 2
 //2.1 Функция для проверки длины строки. Она принимает строку, которую нужно проверить, и максимальную длину и возвращает true, если строка меньше или равна указанной длине, и false, если строка длиннее.
 const stringNotLongerThanLength = (string, length) => string.length <= length;
