+++ js/upload-images.js
@@ -1,0 +1,1 @@
+// Модуль загрузки изображений
