+++ js/util.js
@@ -1,0 +1,13 @@
+// Модуль вспомогательных утилит
+//Функция getRandomInteger возвращает случайное целое число в диапазоне от a до b
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+//Функция getRandomArrayElement возвращает случайноный элемент массива elements
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomInteger, getRandomArrayElement};
