+++ js/view-images.js
@@ -1,0 +1,1 @@
+// Модуль просмотра изображений
